add Cr-cat feature to offers
