+++ packages/wallets/src/components/index.ts
@@ -8,3 +8,5 @@ export { default as WalletStatus } from './WalletStatus';
 export { default as WalletReceiveAddressField } from './WalletReceiveAddressField';
 export { default as WalletConnections } from './WalletConnections';
 export { default as WalletGraphTooltip } from './WalletGraphTooltip';
+export { default as CrCatAuthorizedProviders } from './crCat/CrCatAuthorizedProviders';
+export { default as CrCatFlags } from './crCat/CrCatFlags';
